requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
boto3==1.38.45
botocore==1.38.45
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
daphne==4.1.2
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
docxcompose==1.4.0
docxtpl==0.20.0
drf-nested-routers==0.94.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
lxml==6.0.0
MarkupSafe==3.0.2
model-bakery==1.20.1
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
s3transfer==0.13.0
service-identity==24.2.0
setuptools==80.7.1
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.8.2
zope.interface==7.2
mysqlclient==2.2.7